Data.Text.Case

Description

Most of the functions in this module are subject to fusion, meaning that a pipeline of such functions will usually allocate at most one Text value.

Character manipulation

lowerHead :: Text -> Text

Lowercase the first character of a piece of text.

>>> lowerHead "Title Cased"
"title Cased"

upperHead :: Text -> Text

Uppercase the first character of a piece of text.

>>> upperHead "snake_cased"
"Snake_cased"

mapHead :: (Char -> Char) -> Text -> Text

Apply a function to the first character of a piece of text.

Line manpipulation

indentLines :: Int -> Text -> Text

Indent newlines by the given number of spaces.

prependLines :: Text -> Text -> Text

Prepend newlines with the given separator

Subwords

Removing words

takeWord :: Text -> Text

O(n) Returns the first word, or the original text if no word boundary is encountered. Subject to fusion.

dropWord :: Text -> Text

O(n) Return the suffix after dropping the first word. If no word boundary is encountered, the result will be empty. Subject to fusion.

stripWord :: Text -> Maybe Text

O(n) Return the suffix after removing the first word, or Nothing if no word boundary is encountered.

>>> stripWord "HTML5Spaghetti"
Just "Spaghetti"

>>> stripWord "noboundaries"
Nothing

Breaking on words

breakWord :: Text -> (Text, Text)

Break a piece of text after the first word boundary is encountered.

>>> breakWord "PascalCasedVariable"
("Pacal", "CasedVariable")

>>> breakWord "spinal-cased-variable"
("spinal", "cased-variable")

splitWords :: Text -> [Text]

O(n) Split into a list of words delimited by boundaries.

>>> splitWords "SupercaliFrag_ilistic"
["Supercali","Frag","ilistic"]

Acronyms

toAcronym :: Text -> Maybe Text

O(n) Create an adhoc acronym from a piece of cased text.

>>> toAcronym "AmazonWebServices"
Just "AWS"

>>> toAcronym "Learn-You Some_Haskell"
Just "LYSH"

>>> toAcronym "this_is_all_lowercase"
Nothing

Ordinals

newtype Ordinal a

Constructors

Ordinal a

toOrdinal :: Integral a => a -> Text

Build an ordinal used to denote the position in an ordered sequence. toOrdinal == build . Ordinal.

>>> toOrdinal (101 :: Int)
"101st"

>>> toOrdinal (12 :: Int)
"12th"

Casing

toTitle :: Text -> Text

O(n) Convert casing to Titled Cased Phrase. Subject to fusion.

toCamel :: Text -> Text

O(n) Convert casing to camelCasedPhrase. Subject to fusion.

toPascal :: Text -> Text

O(n) Convert casing to PascalCasePhrase. Subject to fusion.

toSnake :: Text -> Text

O(n) Convert casing to snake_cased_phrase. Subject to fusion.

toSpinal :: Text -> Text

O(n) Convert casing to spinal-cased-phrase. Subject to fusion.

toTrain :: Text -> Text

O(n) Convert casing to Train-Cased-Phrase. Subject to fusion.